Rename EventsInfo file to EventsInfoPublic
